Fixed bugs in character creator, improved maps
- Added ability to drag from character sheet onto map
- Allowed players to drag characters onto the map
- Fixed ability rendering in character creator
- Fixed issue with character creator just looping
- Fixed typo in help text in character creator

+++ frontend/src/windows/Window.ts
@@ -671,7 +671,7 @@ export async function InputDialog(title: string, inputs: { [label: string]: any 
                 continue;
             }
             if (inputElement.type == "number") {
-                value = parseInt(inputElement.value);
+                value = parseFloat(inputElement.value);
             }
             else if (inputElement.type == "checkbox") {
                 value = inputElement.checked;
